Module lea_parser::parser

Wrapper around rust-peg generated parser methods.

Structs

ParseError
Custom error type adding a dummy span and providing a format method.

Functions

block
Parses a block of statements

expression
Parses an expression

ident

literal

parse_expr_as_main
Parses an expression and builds a Function node that will evalute and return the expression when executed.

parse_main
Parses a block of statements and builds a function wrapper that can be run as the main function around it.

statement
Parses a statement