Enum lea_parser::parsetree::StmtKind

pub enum StmtKind<'a> {
    Decl(Vec<Spanned<&'a str>>, Vec<Expr<'a>>),
    Assign(Vec<Variable<'a>>, Vec<Expr<'a>>),
    Do(Block<'a>),
    Break,
    Semi,
    Return(Vec<Expr<'a>>),
    Call(Call<'a>),
    Func(Variable<'a>, Function<'a>),
    Method(Variable<'a>, Spanned<&'a str>, Function<'a>),
    LocalFunc(Spanned<&'a str>, Function<'a>),
    If {
        cond: Expr<'a>,
        body: Block<'a>,
        elifs: Vec<Spanned<(Expr<'a>, Block<'a>)>>,
        el: Option<Block<'a>>,
    },
    While {
        cond: Expr<'a>,
        body: Block<'a>,
    },
    Repeat {
        abort_on: Expr<'a>,
        body: Block<'a>,
    },
    For {
        var: Spanned<&'a str>,
        start: Expr<'a>,
        step: Option<Expr<'a>>,
        end: Expr<'a>,
        body: Block<'a>,
    },
    ForIn {
        vars: Vec<Spanned<&'a str>>,
        iter: Vec<Expr<'a>>,
        body: Block<'a>,
    },
}

Statement nodes

Variants
Decl(Vec<Spanned<&'a str>>, Vec<Expr<'a>>)

Declare a list of locals and assign initial values.

Initial values are optional and default to nil (the second vector can have less elements than the first).

Assign(Vec<Variable<'a>>, Vec<Expr<'a>>)

Assigns a list of expressions to a list of variables.

Contains at least one pair. If the last expression is a function or varargs, all return values are considered for assignment to the leftover variables.

Might contain less variables than expressions or less expressions than variables. In the latter case, the leftover variables are assigned to nil.

Do(Block<'a>)

Execute a block in a new scope

Break

Abort the current loop

Semi

No-op semicolon statement

Return(Vec<Expr<'a>>)

Return a possibly empty list of values to the caller

Call(Call<'a>)

Function call as statement

Func(Variable<'a>, Function<'a>)

Assign function to named variable.

function XY(...) ... end

is equivalent to

XY = function(...) ... end

Method(Variable<'a>, Spanned<&'a str>, Function<'a>)

Method declaration.

function some.thing:methodname(...) ... end

is equivalent to

some.thing.methodname = function(self, ...) ... end

LocalFunc(Spanned<&'a str>, Function<'a>)

Assign function to newly declared local.

local function XY(...) ... end

is equivalent to

local XY; XY = function(...) ... end

If

Executes body if cond is true and the elseifs or else if not.

Fields

cond: Expr<'a>

body: Block<'a>

elifs: Vec<Spanned<(Expr<'a>, Block<'a>)>>

el: Option<Block<'a>>

While

Loops a block while cond is true

Fields

cond: Expr<'a>

body: Block<'a>

Repeat

Loops a block until abort_on is true. The body is executed at least once (abort_on is checked after the body has run).

Fields

abort_on: Expr<'a>

body: Block<'a>

For

Numeric for loop

Fields

var: Spanned<&'a str>

start: Expr<'a>

step: Option<Expr<'a>>

end: Expr<'a>

body: Block<'a>

ForIn

Generic for loop

Fields

vars: Vec<Spanned<&'a str>>

The loop variables, returned by iterator

iter: Vec<Expr<'a>>

Expression list: Iterator function, state, start value, [ignored ...]

body: Block<'a>